requirements.txt
fastapi==0.88.0
Pillow==9.3.0
uvicorn==0.20.0
oss2==2.16.0
requests==2.28.1
python-multipart==0.0.5